Handle "new" gTLDs as fallback for missing 'tld_serv_list' entry
The strategy of keeping a fixed list of "new" gTLDs compiled with the
source code is voided by IANA's policy of updating their list of TLDs
daily. Rather than cache this list (at this point in time approx. 12KB),
assume that any potentially valid new TLD is in this list.

The only satisfying caching strategy is if this 12KB list of TLDs is
cached between executions, but making the library create a persistent
(e.g. file-based) cache between program execution seems bad.

The main drawback of assuming that all TLDs potentially exist is that
it's not easy to differentiate between a network error, being unable to
connect to a presumed WHOIS server of a "new" gTLD, and non-existence of
that WHOIS server.

This commit:
 - changes the formatting of various functions.
 - generalises `splitOn` to handle multiple separators
 - tries to identify syntactically invalid IPv4 / IPv6 / domain names
 - handles WHOIS servers marked with 'VERISIGN' in 'tld_serv_list'
   according to the `whois` program:

   > When querying the Verisign gTLDs (e.g. .com, .net...) thin registry
   > servers for a domain the program will automatically prepend the
   > 'domain ' keyword to only show domain records.

This commit does not handle IDNs, so punycode tests fail.

This commit introduces tests that deliberately fail for `isUnsupported`.